chore: migrate AddExistingTeamView to hooks
